better way of condensing seeds
much fewer losses in O(nlg m) where n is the number of speeds
and m is the sequence length in h.p. condensed alphabet. This
comes at a memory cost of O(m) for keeping the starting positions
(in original alphabet) of all h.p. letters.

Also throws out a bunch of unused code